+++ hugegraph-llm/src/hugegraph_llm/utils/decorators.py
@@ -90,3 +90,14 @@ def record_qps(func: Callable) -> Callable:
             log.debug("%s QPS: %f/s", args[0].__class__.__name__, qps)
         return result
     return wrapper
+
+# A decorator to wrap functions with a task id generation and logging
+def with_task_id(func):
+    def wrapper(*args, **kwargs):
+        import uuid
+        task_id = str(uuid.uuid4())
+        log.info(f"New task created with id: {task_id}")
+        # Optionally, you could also pass the task_id to the function if needed:
+        # kwargs['task_id'] = task_id
+        return func(*args, **kwargs)
+    return wrapper
